🏗️ feat: Introduce Permissions module and refactor role-based access control
