+++ packages/data-provider/src/index.ts
@@ -15,8 +15,8 @@ export * from './models';
 /* mcp */
 export * from './mcp';
 /* RBAC */
+export * from './permissions';
 export * from './roles';
-export * from './groups';
 /* types (exports schemas from `./types` as they contain needed in other defs) */
 export * from './types';
 export * from './types/agents';
