+++ packages/data-provider/src/roles.ts
@@ -1,4 +1,14 @@
 import { z } from 'zod';
+import {
+  Permissions,
+  PermissionTypes,
+  agentPermissionsSchema,
+  promptPermissionsSchema,
+  runCodePermissionsSchema,
+  bookmarkPermissionsSchema,
+  multiConvoPermissionsSchema,
+  temporaryChatPermissionsSchema,
+} from './permissions';
 
 /**
  * Enum for System Defined Roles
@@ -13,79 +23,6 @@ export enum SystemRoles {
    */
   USER = 'USER',
 }
-
-/**
- * Enum for Permission Types
- */
-export enum PermissionTypes {
-  /**
-   * Type for Prompt Permissions
-   */
-  PROMPTS = 'PROMPTS',
-  /**
-   * Type for Bookmark Permissions
-   */
-  BOOKMARKS = 'BOOKMARKS',
-  /**
-   * Type for Agent Permissions
-   */
-  AGENTS = 'AGENTS',
-  /**
-   * Type for Multi-Conversation Permissions
-   */
-  MULTI_CONVO = 'MULTI_CONVO',
-  /**
-   * Type for Temporary Chat
-   */
-  TEMPORARY_CHAT = 'TEMPORARY_CHAT',
-  /**
-   * Type for using the "Run Code" LC Code Interpreter API feature
-   */
-  RUN_CODE = 'RUN_CODE',
-}
-
-/**
- * Enum for Role-Based Access Control Constants
- */
-export enum Permissions {
-  SHARED_GLOBAL = 'SHARED_GLOBAL',
-  USE = 'USE',
-  CREATE = 'CREATE',
-  UPDATE = 'UPDATE',
-  READ = 'READ',
-  READ_AUTHOR = 'READ_AUTHOR',
-  SHARE = 'SHARE',
-}
-
-export const promptPermissionsSchema = z.object({
-  [Permissions.SHARED_GLOBAL]: z.boolean().default(false),
-  [Permissions.USE]: z.boolean().default(true),
-  [Permissions.CREATE]: z.boolean().default(true),
-  // [Permissions.SHARE]: z.boolean().default(false),
-});
-
-export const bookmarkPermissionsSchema = z.object({
-  [Permissions.USE]: z.boolean().default(true),
-});
-
-export const agentPermissionsSchema = z.object({
-  [Permissions.SHARED_GLOBAL]: z.boolean().default(false),
-  [Permissions.USE]: z.boolean().default(true),
-  [Permissions.CREATE]: z.boolean().default(true),
-  // [Permissions.SHARE]: z.boolean().default(false),
-});
-
-export const multiConvoPermissionsSchema = z.object({
-  [Permissions.USE]: z.boolean().default(true),
-});
-
-export const temporaryChatPermissionsSchema = z.object({
-  [Permissions.USE]: z.boolean().default(true),
-});
-
-export const runCodePermissionsSchema = z.object({
-  [Permissions.USE]: z.boolean().default(true),
-});
 
 export const roleSchema = z.object({
   name: z.string(),
